Added a functional taskbar
I still need to change the sizing and theme of the taskbar, as well as create a better dashboard page. I am not sure how to handle those things yet, but I don't want this part to sit for any longer. I will add a new Trello task to fix these things.